refactor(sim): Extract metrics module from main.rs
- Move MetricsEvent enum, log_metrics(), and init_metrics_file() to metrics.rs
- Remove 186 lines of duplicate code from main.rs
- main.rs now imports from the metrics module

This improves code organization and maintainability.
